Use long row counts

+++ h2/src/main/org/h2/command/CommandContainer.java
@@ -60,9 +60,9 @@ public class CommandContainer extends Command {
         }
 
         @Override
-        public int getRowCount() {
+        public long getRowCount() {
             // Not required
-            return 0;
+            return 0L;
         }
 
         @Override
